[add] add bendahara spt

+++ database/migrations/2023_11_08_085131_create_bank_account_table.php
@@ -12,7 +12,7 @@ return new class extends Migration {
     {
         Schema::create('bank_account', function (Blueprint $table) {
             $table->uuid('id')->primary();
-            $table->integer('account_number');
+            $table->bigInteger('account_number');
             $table->timestamps();
         });
     }
